starting the code cleanup. tox_harness.py will contain everything related to handling tox invocation
