--- scripts/devops_tasks/ToxWorkItem.py
@@ -1,9 +1,0 @@
-import logging
-import os
-from common_tasks import process_glob_string, run_check_call, cleanup_folder
-
-class ToxWorkItem:
-    def __init__(self, target_package_path, tox_env, options_array):
-        self.target_package_path = target_package_path
-        self.tox_env = tox_env
-        self.options_array = options_array
